refactor(fleet-change): extract rule validation
